# api/openapi.yaml
openapi: 3.1.0
info:
  title: uspace-ansp national API
  version: 1.0.0
  description: |
    The U-space interface of the air navigation service provider in the
    Georgian U-space system-of-systems (Reg. (EU) 2021/664, 2021/665):
    dynamic airspace reconfiguration (ATS.TR.237), the manned traffic
    information service (ATS.OR.127), the Annex V coordination inbox
    (Art. 13(2)) and the accounts of the dispatcher console.

    This file is the contract (CLAUDE.md rule 7): an endpoint that is
    not in it does not exist; the server interface, the models and the
    Go client in api/gen are generated from it and committed, and CI
    fails on a stale generation. Changes within /v1 are additive only
    (spec 00 §7, 02 §1): new optional fields, new operations, new enum
    values announced one minor ahead; a breaking change is a /v2 served
    in parallel. Inbound bodies leave additionalProperties open: unknown
    members are ignored, never refused (02 §1).

    Conventions (02 §1): every time is RFC 3339 UTC with Z and
    millisecond precision; every geometry is GeoJSON in WGS84 (SRID
    4326) with positions [lng, lat]; every distance is in metres and
    every altitude is a number in metres with its reference in a _ref
    member (AMSL or WGS84; AGL is refused for a dynamic restriction in
    this release, D3). Every error is application/problem+json in the
    one body of the ecosystem (Problem, M28), and every 429 and 503
    carries Retry-After.

...
  license:
...
servers:
...
components:
  securitySchemes:
    ecosystemToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An ecosystem RS256 token from the authority's token service (or
        the lab issuer), verified by uspace-core/auth: iss on
        ANSP_TOKEN_ISSUERS, aud a host of ANSP_AUDIENCES (M18), exp with
        30 s skew, the scopes the operation names.
    consoleSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A console session issued by this system at POST /v1/auth/mfa
        (scope session, roles [role], realm console, aud this system's
        host; M20), forwarded by the BFF as a bearer.
    sessionCookie:
      type: apiKey
      in: cookie
      name: uspace_session
      description: |
        The same session in the uspace_session cookie, accepted only on a
        same-origin WebSocket upgrade whose Origin is on
        ANSP_WS_ALLOWED_ORIGINS (M22).
...
paths:
  /v1/restrictions:
    post:
      operationId: createRestriction
      tags: [restrictions]
      summary: Plan a restriction
      x-process: api
      x-spec: 02 §3 ansp, 02 F2, 01 N1
      x-auth: session:watch_supervisor
      description: |
        Plans a dynamic restriction (state planned) inside a designated
        U-space airspace (2021/665 ATS.TR.237(a); 2021/664 Art. 4). The
        restriction must intersect a current USPACE feature of the CIS
        projection and name it in uspace_airspace_id (M9); with a stale
        projection the answer is 503 cis_stale. The window and the area
        are checked against the F3548 constraint limits of uspace-core
        (CstrMaxDurationHours, CstrMaxPlanningHorizonDays, CstrMaxAreaKm2,
        CstrMaxVertices). Vertical limits are in AMSL or WGS84; AGL is
        refused with "lower_ref: AGL is not supported for a dynamic
        restriction in this release" (D3). Idempotency-Key is the
        client's reference: a repeat with the same key and body answers
        200 with the restriction first created; with another body, 409.
      security:
        - consoleSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestrictionCreate"
            example:
              uspace_airspace_id: GEOTU01
              zone_type: PROHIBITED
              geometry:
                type: Polygon
                coordinates: [[[44.78, 41.70], [44.82, 41.70], [44.82, 41.73], [44.78, 41.73], [44.78, 41.70]]]
              lower_m: 0
              lower_ref: AMSL
              upper_m: 1200
              upper_ref: AMSL
              starts_at: "2026-10-02T12:00:00.000Z"
              ends_at: "2026-10-02T16:00:00.000Z"
              reason_text: Search and rescue operation (synthetic example)
      responses:
        "201":
          description: Planned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restriction"
              example: *RestrictionPlanned
        "200":
          description: A repeat of an earlier request with the same Idempotency-Key and body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restriction"
              example: *RestrictionPlanned
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/TooLarge"
        "503":
          $ref: "#/components/responses/Unavailable"
    get:
      operationId: listRestrictions
      tags: [restrictions]
      summary: List restrictions
      x-process: api
      x-spec: 02 §3 ansp
      x-auth: token:ansp.coordination or session
      description: |
        The restrictions with their states and delivery status
        (ATS.TR.237(b)), newest first, filtered by state, by the instant
        at which they apply and by a bounding box. Answers carry the CIS
        version the restrictions were placed against and its age.
      security:
        - consoleSession: []
        - ecosystemToken: [ansp.coordination]
      parameters:
        - name: state
          in: query
          required: false
          description: Only restrictions in this state.
          schema:
            $ref: "#/components/schemas/RestrictionState"
          example: active
        - name: at
          in: query
          required: false
          description: Only restrictions whose window contains this instant.
          schema:
            type: string
            format: date-time
          example: "2026-10-02T13:00:00.000Z"
        - $ref: "#/components/parameters/BBox"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The restrictions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestrictionList"
              example:
                restrictions:
                  - *RestrictionActive
                cis_version: "42"
                cis_age_s: 12.5
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/restrictions/{id}:
    get:
      operationId: getRestriction
      tags: [restrictions]
      summary: One restriction
      x-process: api
      x-spec: 02 §3 ansp
      x-auth: token:ansp.coordination or session
      description: |
        The restriction, its current ED-318 feature, its F3548 constraint
        reference and its deliveries summary (ATS.TR.237(b)). This is
        also the pull_url of a degraded direct delivery (02 F2 failure
        rule, M5).
      security:
        - consoleSession: []
        - ecosystemToken: [ansp.coordination]
      parameters:
        - $ref: "#/components/parameters/RestrictionID"
      responses:
        "200":
          description: The restriction.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restriction"
              example: *RestrictionActive
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/restrictions/{id}/versions:
    get:
      operationId: listRestrictionVersions
      tags: [restrictions]
      summary: Every version of a restriction
      x-process: api
      x-spec: 01 N4, 02 §3 ansp
      x-auth: session
      description: |
        Every version as published, oldest first: who changed it, when
        and why, until when (Annex V(1), the N4 record of every
        reconfiguration).
      security:
        - consoleSession: []
      parameters:
        - $ref: "#/components/parameters/RestrictionID"
      responses:
        "200":
          description: The versions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestrictionVersionList"
              example:
                versions:
                  - *RestrictionVersion
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/restrictions/{id}/versions/{version}:
    get:
      operationId: getRestrictionVersion
      tags: [restrictions]
      summary: One version of a restriction
      x-process: api
      x-spec: 01 N4, 02 §3 ansp
      x-auth: session
      description: The version as published (Annex V(1)).
      security:
        - consoleSession: []
      parameters:
        - $ref: "#/components/parameters/RestrictionID"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: The version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestrictionVersion"
              example: *RestrictionVersion
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/restrictions/{id}/activate:
    post:
      operationId: activateRestriction
      tags: [restrictions]
      summary: Activate a planned restriction
      x-process: api
      x-spec: ATS.TR.237(b) activation
      x-auth: session:watch_supervisor
      description: |
        planned → active, at once or at starts_at when that is in the
        future (2021/665 ATS.TR.237(b): activation). Starts the outbox:
        the CISP publication and the DSS constraint are written in
        parallel (D6), each idempotent on (ansp_ref, ansp_version).
      security:
        - consoleSession: []
      parameters:
        - $ref: "#/components/parameters/RestrictionID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonRequest"
            example:
              reason: Rescue helicopter on scene
      responses:
        "200":
          description: Active (or scheduled to activate at starts_at).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restriction"
              example: *RestrictionActive
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/restrictions/{id}/extend:
    post:
      operationId: extendRestriction
      tags: [restrictions]
      summary: Extend a restriction
      x-process: api
      x-spec: ATS.TR.237(b) temporary limitation
      x-auth: session:watch_supervisor
      description: |
        A new ends_at (2021/665 ATS.TR.237(b): temporary limitation).
        Within CstrMaxDurationHours of starts_at the restriction gets a
        new ansp_version; beyond it a re-issue is created and linked by
        supersedes_id, and the answer is that re-issue.
      security:
        - consoleSession: []
      parameters:
        - $ref: "#/components/parameters/RestrictionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtendRequest"
            example:
              ends_at: "2026-10-02T18:00:00.000Z"
              reason: Operation continues
      responses:
        "200":
          description: Extended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restriction"
              example: *RestrictionActive
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/restrictions/{id}/end:
    post:
      operationId: endRestriction
      tags: [restrictions]
      summary: End an active restriction now
      x-process: api
      x-spec: ATS.TR.237(b) deactivation
      x-auth: session:watch_supervisor
      description: "active → ended now (2021/665 ATS.TR.237(b): deactivation)."
      security:
        - consoleSession: []
      parameters:
        - $ref: "#/components/parameters/RestrictionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonRequest"
            example:
              reason: Rescue completed
      responses:
        "200":
          description: Ended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restriction"
              example: *RestrictionEnded
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/restriction-requests/{id}:
    get:
      operationId: getRestrictionRequest
      tags: [restriction-requests]
      summary: The state of a restriction request
      x-process: api
      x-spec: 02 F11
      x-auth: token:ansp.requests or session
      description: received, accepted (with restriction_id) or declined (with decision_reason) (02 F11).
      security:
        - ecosystemToken: [ansp.requests]
        - consoleSession: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "200":
          description: The request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestrictionRequest"
              example: *RestrictionRequestAccepted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...